rearrange some logic now that it's all in place
