+++ dashboard/app/models/lesson.rb
@@ -227,21 +227,10 @@ class Lesson < ApplicationRecord
 
   def lesson_plan_pdf_url
     if script.is_migrated && has_lesson_plan
-      "https://cdo-lesson-plans.s3.amazonaws.com/#{lesson_plan_pdf_pathname}" if lesson_plan_pdf_pathname.present?
+      Services::LessonPlanPdfs.get_url(self)
     else
       "#{lesson_plan_base_url}/Teacher.pdf"
     end
-  end
-
-  def lesson_plan_pdf_pathname
-    version_number =
-      begin
-        Time.parse(script.seeded_at).to_s(:number)
-      rescue ArgumentError, TypeError
-        return nil
-      end
-    filename = ActiveStorage::Filename.new(key + ".pdf").sanitized
-    return Pathname.new(File.join(script.name, version_number, filename))
   end
 
   def lesson_plan_base_url
