+++ dashboard/lib/services/script_seed.rb
@@ -12,7 +12,7 @@
 
 module Services
   module ScriptSeed
-    prepend PdfGeneration
+    prepend LessonPlanPdfs::ScriptSeed
 
     # Holds data that we've already retrieved from the database. Used to look up
     # associations of objects without making additional queries.
